agents/timezone-arbitrage/requirements.txt
# 时区套利 Agent 依赖

# 核心
pyyaml>=6.0
aiohttp>=3.9.0
asyncio>=3.4.3

# 时间处理
pytz>=2024.1
python-dateutil>=2.8.2

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# HTTP 请求
requests>=2.31.0
feedparser>=6.0.10  # RSS 解析

# Polymarket SDK
polymarket-python-sdk>=1.0.0

# 数据库
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# 缓存
redis>=5.0.0

# 日志
python-json-logger>=2.0.7

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 开发
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
